Move src/core/lib to src/core
