+++ src/core/index.js
@@ -1,13 +1,13 @@
-var Resource = require('./lib/resources/resource')
-var MessageList = require('./lib/resources/message_list')
-var Presence = require('./lib/resources/presence')
-var Status = require('./lib/resources/status')
-var Stream = require('./lib/resources/stream')
+var Resource = require('./resources/resource')
+var MessageList = require('./resources/message_list')
+var Presence = require('./resources/presence')
+var Status = require('./resources/status')
+var Stream = require('./resources/stream')
 
 module.exports = {
   Persistence: require('persistence'),
-  Type: require('./lib/type.js'),
-  PresenceManager: require('./lib/resources/presence/presence_manager.js'),
+  Type: require('./type.js'),
+  PresenceManager: require('./resources/presence/presence_manager.js'),
 
   Resource: Resource,
   MessageList: MessageList,
