+++ src/core/resources/message_list/index.js
