+++ src/core/resources/presence/presence_manager.js
