+++ src/core/resources/presence/presence_store.js
