+++ src/core/resources/presence/readme.md
