+++ src/core/resources/stream/index.js
