+++ src/core/resources/stream/subscriber_state.js
