+++ src/core/type.js
