+++ test/client.presence.remote.test.js
@@ -3,7 +3,7 @@ var common = require('./common.js')
 var assert = require('assert')
 var Persistence = require('../src/core').Persistence
 var Tracker = require('callback_tracker')
-var PresenceManager = require('../src/core/lib/resources/presence/presence_manager.js')
+var PresenceManager = require('../src/core/resources/presence/presence_manager.js')
 var AssertHelper = require('./lib/assert_helper.js')
 var PresenceMessage = AssertHelper.PresenceMessage
 var presenceManagerForSentry = AssertHelper.presenceManagerForSentry
